@@ -25,9 +25,16 @@ package org.shadowice.flocke.andotp.View;
 import android.animation.ObjectAnimator;
 import android.content.Context;
 import android.graphics.ColorFilter;
+
+import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
+
+import android.graphics.Typeface;
+import android.text.Spannable;
+import android.text.SpannableStringBuilder;
 import android.text.TextUtils;
+import android.text.style.StyleSpan;
 import android.view.View;
 import android.view.animation.LinearInterpolator;
 import android.widget.FrameLayout;
@@ -38,6 +45,7 @@ import android.widget.TextView;
 
 import org.shadowice.flocke.andotp.Database.Entry;
 import org.shadowice.flocke.andotp.R;
+import org.shadowice.flocke.andotp.Utilities.Constants;
 import org.shadowice.flocke.andotp.Utilities.EntryThumbnail;
 import org.shadowice.flocke.andotp.Utilities.Settings;
 import org.shadowice.flocke.andotp.Utilities.Tools;
@@ -52,26 +60,24 @@ import static org.shadowice.flocke.andotp.Activities.MainActivity.animatorDurati
 
 public class EntryViewHolder extends RecyclerView.ViewHolder
         implements ItemTouchHelperViewHolder {
-    private Context context;
+    private final Context context;
     private Callback callback;
     private boolean tapToReveal;
 
-    private CardView card;
-    private LinearLayout valueLayout;
-    private LinearLayout coverLayout;
-    private LinearLayout counterLayout;
-    private FrameLayout thumbnailFrame;
-    private ImageView visibleImg;
-    private ImageView thumbnailImg;
-    private ImageButton menuButton;
-    private ImageButton copyButton;
-    private TextView value;
-    private TextView issuer;
-    private TextView label;
-    private TextView separator;
-    private TextView counter;
-    private TextView tags;
-    private MaterialProgressBar progressBar;
+    private final CardView card;
+    private final LinearLayout valueLayout;
+    private final LinearLayout coverLayout;
+    private final LinearLayout counterLayout;
+    private final FrameLayout thumbnailFrame;
+    private final ImageView visibleImg;
+    private final ImageView thumbnailImg;
+    private final ImageButton menuButton;
+    private final ImageButton copyButton;
+    private final TextView value;
+    private final TextView label;
+    private final TextView counter;
+    private final TextView tags;
+    private final MaterialProgressBar progressBar;
 
     public EntryViewHolder(Context context, final View v, boolean tapToReveal) {
         super(v);
@@ -85,9 +91,7 @@ public class EntryViewHolder extends RecyclerView.ViewHolder
         thumbnailFrame = v.findViewById(R.id.thumbnailFrame);
         thumbnailImg = v.findViewById(R.id.thumbnailImg);
         coverLayout = v.findViewById(R.id.coverLayout);
-        issuer = v.findViewById(R.id.textViewIssuer);
         label = v.findViewById(R.id.textViewLabel);
-        separator = v.findViewById(R.id.textViewSeparator);
         tags = v.findViewById(R.id.textViewTags);
         counterLayout = v.findViewById(R.id.counterLayout);
         counter = v.findViewById(R.id.counter);
@@ -105,55 +109,49 @@ public class EntryViewHolder extends RecyclerView.ViewHolder
         visibleImg.getDrawable().setColorFilter(colorFilter);
         invisibleImg.getDrawable().setColorFilter(colorFilter);
 
-        // Setup onClickListeners
-        menuButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (callback != null)
-                    callback.onMenuButtonClicked(view, getAdapterPosition());
-            }
-        });
+        setupOnClickListeners(menuButton, copyButton);
 
-        copyButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (callback != null)
-                    callback.onCopyButtonClicked(value.getTag().toString(), getAdapterPosition());
-            }
-        });
+        setTapToReveal(tapToReveal);
+    }
 
-        counterLayout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (callback != null)
-                    callback.onCounterClicked(getAdapterPosition());
-            }
-        });
+    private void setupOnClickListeners(ImageButton menuButton, ImageButton copyButton) {
+        menuButton.setOnClickListener(view ->
+            adapterPositionSafeCallback((callback, adapterPosition) ->
+                callback.onMenuButtonClicked(view, adapterPosition)
+            )
+        );
 
-        counterLayout.setOnLongClickListener(new View.OnLongClickListener() {
-            @Override
-            public boolean onLongClick(View view) {
-                if (callback != null)
-                    callback.onCounterLongPressed(getAdapterPosition());
+        copyButton.setOnClickListener(view ->
+            adapterPositionSafeCallback((callback, adapterPosition) ->
+                callback.onCopyButtonClicked(value.getTag().toString(), adapterPosition)
+            )
+        );
 
-                return false;
-            }
+        counterLayout.setOnClickListener(view ->
+            adapterPositionSafeCallback(Callback::onCounterClicked)
+        );
+
+        counterLayout.setOnLongClickListener(view -> {
+            adapterPositionSafeCallback(Callback::onCounterLongPressed);
+            return false;
         });
 
         card.setOnClickListener(new SimpleDoubleClickListener() {
             @Override
             public void onSingleClick(View v) {
-                if (callback != null)
-                    callback.onCardSingleClicked(getAdapterPosition(), value.getTag().toString());
+                adapterPositionSafeCallback((callback, adapterPosition) ->
+                    callback.onCardSingleClicked(adapterPosition, value.getTag().toString())
+                );
             }
 
             @Override
             public void onDoubleClick(View v) {
-                if (callback != null)
-                    callback.onCardDoubleClicked(getAdapterPosition(), value.getTag().toString());
+                adapterPositionSafeCallback((callback, adapterPosition) ->
+                    callback.onCardDoubleClicked(adapterPosition, value.getTag().toString())
+                );
             }
         });
-
+        
         itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -161,8 +159,20 @@ public class EntryViewHolder extends RecyclerView.ViewHolder
                     callback.onItemClickListener(getAdapterPosition());
             }
         });
+    }
 
-        setTapToReveal(tapToReveal);
+    @FunctionalInterface
+    private interface AdapterPositionSafeCallbackConsumer {
+        /** The specified {@link Callback} is guaranteed to be non-null, and adapterPosition is
+         * guaranteed to be a valid position. */
+        void accept(@NonNull Callback callback, int adapterPosition);
+    }
+
+    private void adapterPositionSafeCallback(AdapterPositionSafeCallbackConsumer safeCallback) {
+        int clickedPosition = getAdapterPosition();
+        if (callback != null && clickedPosition != RecyclerView.NO_POSITION) {
+            safeCallback.accept(callback, clickedPosition);
+        }
     }
 
     public void updateValues(Entry entry) {
@@ -177,35 +187,40 @@ public class EntryViewHolder extends RecyclerView.ViewHolder
 
         final String tokenFormatted = Tools.formatToken(entry.getCurrentOTP(), settings.getTokenSplitGroupSize());
 
+        String issuerText = entry.getIssuer();
+        String labelText = entry.getLabel();
+
         String contentHint = "";
 
-        String issuerText = entry.getIssuer();
-        if (!TextUtils.isEmpty(issuerText)) {
-            issuer.setText(issuerText);
-            issuer.setVisibility(View.VISIBLE);
+        SpannableStringBuilder labelBuilder = new SpannableStringBuilder();
+
+        if (!TextUtils.isEmpty(issuerText) && !settings.isHideIssuerEnabled()) {
+            labelBuilder.append(issuerText);
+
+            labelBuilder.setSpan(new StyleSpan(Typeface.BOLD), 0, issuerText.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
 
             contentHint = issuerText;
-        } else {
-            issuer.setVisibility(View.GONE);
         }
 
-        String labelText = entry.getLabel();
-        if (!TextUtils.isEmpty(labelText)) {
-            label.setText(labelText);
-            label.setVisibility(View.VISIBLE);
+        if (!TextUtils.isEmpty(issuerText) && !TextUtils.isEmpty(labelText) && !settings.isHideIssuerEnabled()) {
+            String separatorText = "\u00a0-\u00a0"; // \u00a0 = non-breaking space
 
-            contentHint = labelText;
-        } else {
-            label.setVisibility(View.GONE);
-        }
+            if (settings.getCardLayout() == Constants.CardLayouts.FULL)
+                separatorText = "\n";
 
-        if (!TextUtils.isEmpty(issuerText) && !TextUtils.isEmpty(labelText)) {
-            separator.setVisibility(View.VISIBLE);
+            labelBuilder.append(separatorText);
 
             contentHint = issuerText + " - " + labelText;
-        } else {
-            separator.setVisibility(View.GONE);
         }
+
+        if (!TextUtils.isEmpty(labelText)) {
+            labelBuilder.append(labelText);
+
+            if (TextUtils.isEmpty(issuerText) || settings.isHideIssuerEnabled())
+                contentHint = labelText;
+        }
+
+        label.setText(labelBuilder);
 
         copyButton.setContentDescription(context.getString(R.string.button_card_copy_format, contentHint));
         menuButton.setContentDescription(context.getString(R.string.button_card_options_format, contentHint));
@@ -280,15 +295,26 @@ public class EntryViewHolder extends RecyclerView.ViewHolder
         thumbnailImg.requestLayout();
     }
 
-    public void setLabelScroll(boolean active) {
-        if (active) {
-            label.setEllipsize(TextUtils.TruncateAt.MARQUEE);
-            label.setHorizontallyScrolling(true);
-            label.setSelected(true);
-        } else {
-            label.setEllipsize(TextUtils.TruncateAt.END);
-            label.setHorizontallyScrolling(false);
-            label.setSelected(false);
+    public void setLabelScroll(Constants.LabelDisplay labelDisplay, Constants.CardLayouts cardLayout) {
+        switch (labelDisplay) {
+            case TRUNCATE:
+                label.setEllipsize(TextUtils.TruncateAt.END);
+                label.setHorizontallyScrolling(false);
+                label.setSelected(false);
+                label.setMaxLines(cardLayout == Constants.CardLayouts.FULL ? 2 : 1);
+                break;
+            case SCROLL:
+                label.setEllipsize(TextUtils.TruncateAt.MARQUEE);
+                label.setHorizontallyScrolling(true);
+                label.setSelected(true);
+                label.setMaxLines(cardLayout == Constants.CardLayouts.FULL ? 2 : 1);
+                break;
+            case MULTILINE:
+                label.setEllipsize(null);
+                label.setHorizontallyScrolling(false);
+                label.setSelected(false);
+                label.setMaxLines(10);
+                break;
         }
     }
 
