@@ -24,7 +24,6 @@ package org.shadowice.flocke.andotp.Activities;
 
 import android.app.backup.BackupManager;
 import android.app.AlertDialog;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.net.Uri;
@@ -37,9 +36,12 @@ import android.preference.Preference;
 import android.preference.PreferenceCategory;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
+
+import androidx.annotation.NonNull;
 import androidx.appcompat.widget.Toolbar;
 
 import android.provider.DocumentsContract;
+import android.util.Log;
 import android.view.ViewStub;
 import android.widget.Toast;
 
@@ -66,10 +68,12 @@ import static org.shadowice.flocke.andotp.Utilities.Constants.EncryptionType;
 
 public class SettingsActivity extends BaseActivity
         implements SharedPreferences.OnSharedPreferenceChangeListener{
-    SettingsFragment fragment;
 
-    SecretKey encryptionKey = null;
-    boolean encryptionChanged = false;
+    private SettingsFragment fragment;
+    private SharedPreferences prefs;
+
+    private SecretKey encryptionKey = null;
+    private boolean encryptionChanged = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -104,12 +108,12 @@ public class SettingsActivity extends BaseActivity
                 .replace(R.id.container_content, fragment)
                 .commit();
 
-        SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(this);
-        sharedPref.registerOnSharedPreferenceChangeListener(this);
+        prefs = PreferenceManager.getDefaultSharedPreferences(this);
+        prefs.registerOnSharedPreferenceChangeListener(this);
     }
 
     @Override
-    protected void onSaveInstanceState(Bundle outState) {
+    protected void onSaveInstanceState(@NonNull Bundle outState) {
         super.onSaveInstanceState(outState);
 
         outState.putBoolean(Constants.EXTRA_SETTINGS_ENCRYPTION_CHANGED, encryptionChanged);
@@ -144,9 +148,11 @@ public class SettingsActivity extends BaseActivity
         backupManager.dataChanged();
 
         if (key.equals(getString(R.string.settings_key_theme)) ||
-                key.equals(getString(R.string.settings_key_locale)) ||
+                key.equals(getString(R.string.settings_key_lang)) ||
                 key.equals(getString(R.string.settings_key_special_features)) ||
-                key.equals(getString(R.string.settings_key_backup_location))) {
+                key.equals(getString(R.string.settings_key_backup_location)) ||
+                key.equals(getString(R.string.settings_key_theme_mode)) ||
+                key.equals(getString(R.string.settings_key_theme_black_auto))) {
             recreate();
         } else if(key.equals(getString(R.string.settings_key_encryption))) {
             if (settings.getEncryption() != EncryptionType.PASSWORD) {
@@ -166,6 +172,13 @@ public class SettingsActivity extends BaseActivity
                 if (fragment.useAndroidSync != null)
                     fragment.useAndroidSync.setEnabled(true);
             }
+        } else if(key.equals(getString(R.string.settings_key_enable_android_backup_service)))
+        {
+            Log.d(SettingsActivity.class.getSimpleName(), "onSharedPreferenceChanged called modifying settings_key_enable_android_backup_service service is now: " +
+                    (settings.getAndroidBackupServiceEnabled() ? "enabled" : "disabled"));
+
+            int message = settings.getAndroidBackupServiceEnabled() ? R.string.settings_toast_android_sync_enabled : R.string.settings_toast_android_sync_disabled;
+            Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
         }
 
         fragment.updateAutoBackup();
@@ -271,42 +284,45 @@ public class SettingsActivity extends BaseActivity
                 getContentResolver().takePersistableUriPermission(treeUri, takeFlags);
                 settings.setBackupLocation(treeUri);
             }
-        } else if (fragment.pgpSigningKey.handleOnActivityResult(requestCode, resultCode, data)) {
-            // handled by OpenPgpKeyPreference
-            return;
+        } else {
+            // Handled in OpenPgpKeyPreference
+            fragment.pgpSigningKey.handleOnActivityResult(requestCode, resultCode, data);
         }
     }
 
+    @Override
+    protected void onDestroy() {
+        prefs.unregisterOnSharedPreferenceChangeListener(this);
+        prefs = null;
+
+        super.onDestroy();
+    }
+
     public static class SettingsFragment extends PreferenceFragment {
-        PreferenceCategory catSecurity;
+        PreferenceCategory catUI;
 
-        Settings settings;
-        ListPreference encryption;
-        Preference backupLocation;
-        ListPreference useAutoBackup;
-        CheckBoxPreference useAndroidSync;
+        private Settings settings;
+        private ListPreference encryption;
+        private ListPreference useAutoBackup;
+        private CheckBoxPreference useAndroidSync;
 
-        OpenPgpAppPreference pgpProvider;
-        EditTextPreference pgpEncryptionKey;
-        OpenPgpKeyPreference pgpSigningKey;
+        private EditTextPreference pgpEncryptionKey;
+        private OpenPgpKeyPreference pgpSigningKey;
+        ListPreference themeMode;
+        CheckBoxPreference themeBlack;
+        ListPreference theme;
 
         public void encryptionChangeWithDialog(final EncryptionType encryptionType) {
             AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
             builder.setTitle(R.string.settings_dialog_title_warning)
                     .setMessage(R.string.settings_dialog_msg_encryption_change)
-                    .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialogInterface, int i) {
-                            if (encryptionType == EncryptionType.PASSWORD)
-                                ((SettingsActivity) getActivity()).tryEncryptionChangeWithAuth(encryptionType);
-                            else if (encryptionType == EncryptionType.KEYSTORE)
-                                ((SettingsActivity) getActivity()).tryEncryptionChange(encryptionType, null);
-                        }
+                    .setPositiveButton(android.R.string.ok, (dialogInterface, i) -> {
+                        if (encryptionType == EncryptionType.PASSWORD)
+                            ((SettingsActivity) getActivity()).tryEncryptionChangeWithAuth(encryptionType);
+                        else if (encryptionType == EncryptionType.KEYSTORE)
+                            ((SettingsActivity) getActivity()).tryEncryptionChange(encryptionType, null);
                     })
-                    .setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialogInterface, int i) {
-                        }
+                    .setNegativeButton(android.R.string.cancel, (dialogInterface, i) -> {
                     })
                     .create()
                     .show();
@@ -337,58 +353,58 @@ public class SettingsActivity extends BaseActivity
             addPreferencesFromResource(R.xml.preferences);
 
             CredentialsPreference credentialsPreference = (CredentialsPreference) findPreference(getString(R.string.settings_key_auth));
-            credentialsPreference.setEncryptionChangeCallback(new CredentialsPreference.EncryptionChangeCallback() {
-                @Override
-                public boolean testEncryptionChange(byte[] newKey) {
-                    return ((SettingsActivity) getActivity()).tryEncryptionChange(settings.getEncryption(), newKey);
-                }
-            });
+            credentialsPreference.setEncryptionChangeCallback(newKey -> ((SettingsActivity) getActivity()).tryEncryptionChange(settings.getEncryption(), newKey));
 
             CheckBoxPreference blockAutofill = (CheckBoxPreference) findPreference(getString(R.string.settings_key_block_autofill));
+            CheckBoxPreference autoUnlockAfterAutofill = (CheckBoxPreference) findPreference(getString(R.string.settings_key_auto_unlock_after_autofill));
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 blockAutofill.setEnabled(true);
                 blockAutofill.setSummary(R.string.settings_desc_block_autofill);
+                autoUnlockAfterAutofill.setEnabled(true);
+                autoUnlockAfterAutofill.setSummary(R.string.settings_desc_auto_unlock_after_autofill);
             } else {
                 blockAutofill.setEnabled(false);
-                blockAutofill.setSummary(R.string.settings_desc_block_autofill_android);
+                blockAutofill.setSummary(R.string.settings_desc_autofill_requires_android_o);
+                autoUnlockAfterAutofill.setEnabled(false);
+                autoUnlockAfterAutofill.setSummary(R.string.settings_desc_autofill_requires_android_o);
             }
 
             // Authentication
-            catSecurity = (PreferenceCategory) findPreference(getString(R.string.settings_key_cat_security));
+            catUI = (PreferenceCategory) findPreference(getString(R.string.settings_key_cat_ui));
             encryption = (ListPreference) findPreference(getString(R.string.settings_key_encryption));
+            themeMode = (ListPreference) findPreference(getString(R.string.settings_key_theme_mode));
+            themeBlack = (CheckBoxPreference) findPreference(getString(R.string.settings_key_theme_black_auto));
+            theme = (ListPreference) findPreference(getString(R.string.settings_key_theme));
 
-            encryption.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
-                @Override
-                public boolean onPreferenceChange(final Preference preference, Object o) {
-                    String newEncryption = (String) o;
-                    EncryptionType encryptionType = EncryptionType.valueOf(newEncryption.toUpperCase());
-                    EncryptionType oldEncryptionType = settings.getEncryption();
-                    AuthMethod authMethod = settings.getAuthMethod();
+            encryption.setOnPreferenceChangeListener((preference, o) -> {
+                String newEncryption = (String) o;
+                EncryptionType encryptionType = EncryptionType.valueOf(newEncryption.toUpperCase());
+                EncryptionType oldEncryptionType = settings.getEncryption();
+                AuthMethod authMethod = settings.getAuthMethod();
 
-                    if (encryptionType != oldEncryptionType) {
-                        if (encryptionType == EncryptionType.PASSWORD) {
-                            if (authMethod != AuthMethod.PASSWORD && authMethod != AuthMethod.PIN) {
-                                UIHelper.showGenericDialog(getActivity(), R.string.settings_dialog_title_error, R.string.settings_dialog_msg_encryption_invalid_with_auth);
+                if (encryptionType != oldEncryptionType) {
+                    if (encryptionType == EncryptionType.PASSWORD) {
+                        if (authMethod != AuthMethod.PASSWORD && authMethod != AuthMethod.PIN) {
+                            UIHelper.showGenericDialog(getActivity(), R.string.settings_dialog_title_error, R.string.settings_dialog_msg_encryption_invalid_with_auth);
+                            return false;
+                        } else {
+                            if (settings.getAuthCredentials().isEmpty()) {
+                                UIHelper.showGenericDialog(getActivity(), R.string.settings_dialog_title_error, R.string.settings_dialog_msg_encryption_invalid_without_credentials);
                                 return false;
-                            } else {
-                                if (settings.getAuthCredentials().isEmpty()) {
-                                    UIHelper.showGenericDialog(getActivity(), R.string.settings_dialog_title_error, R.string.settings_dialog_msg_encryption_invalid_without_credentials);
-                                    return false;
-                                }
                             }
-
-                            encryptionChangeWithDialog(EncryptionType.PASSWORD);
-                        } else if (encryptionType == EncryptionType.KEYSTORE) {
-                            encryptionChangeWithDialog(EncryptionType.KEYSTORE);
                         }
-                    }
 
-                    return false;
+                        encryptionChangeWithDialog(EncryptionType.PASSWORD);
+                    } else if (encryptionType == EncryptionType.KEYSTORE) {
+                        encryptionChangeWithDialog(EncryptionType.KEYSTORE);
+                    }
                 }
+
+                return false;
             });
 
             // Backup location
-            backupLocation = findPreference(getString(R.string.settings_key_backup_location));
+            Preference backupLocation = findPreference(getString(R.string.settings_key_backup_location));
 
             if (settings.isBackupLocationSet()) {
                 backupLocation.setSummary(R.string.settings_desc_backup_location_set);
@@ -396,40 +412,25 @@ public class SettingsActivity extends BaseActivity
                 backupLocation.setSummary(R.string.settings_desc_backup_location);
             }
 
-            backupLocation.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                @Override
-                public boolean onPreferenceClick(Preference preference) {
-                    ((SettingsActivity) getActivity()).requestBackupAccess();
-                    return true;
-                }
+            backupLocation.setOnPreferenceClickListener(preference -> {
+                ((SettingsActivity) getActivity()).requestBackupAccess();
+                return true;
             });
 
             // OpenPGP
-            pgpProvider = (OpenPgpAppPreference) findPreference(getString(R.string.settings_key_openpgp_provider));
+            OpenPgpAppPreference pgpProvider = (OpenPgpAppPreference) findPreference(getString(R.string.settings_key_openpgp_provider));
             pgpEncryptionKey = (EditTextPreference) findPreference(getString(R.string.settings_key_openpgp_key_encrypt));
             pgpSigningKey = (OpenPgpKeyPreference) findPreference(getString(R.string.settings_key_openpgp_key_sign));
 
             pgpSigningKey.setOpenPgpProvider(pgpProvider.getValue());
 
-            if (pgpProvider.getValue() != null && ! pgpProvider.getValue().isEmpty()) {
-                pgpEncryptionKey.setEnabled(true);
-            } else {
-                pgpEncryptionKey.setEnabled(false);
-            }
+            pgpEncryptionKey.setEnabled(pgpProvider.getValue() != null && !pgpProvider.getValue().isEmpty());
 
-            pgpProvider.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
-                @Override
-                public boolean onPreferenceChange(Preference preference, Object newValue) {
-                    if (newValue != null && ! ((String) newValue).isEmpty()) {
-                        pgpEncryptionKey.setEnabled(true);
-                    } else {
-                        pgpEncryptionKey.setEnabled(false);
-                    }
+            pgpProvider.setOnPreferenceChangeListener((preference, newValue) -> {
+                pgpEncryptionKey.setEnabled(newValue != null && !((String) newValue).isEmpty());
+                pgpSigningKey.setOpenPgpProvider((String) newValue);
 
-                    pgpSigningKey.setOpenPgpProvider((String) newValue);
-
-                    return true;
-                }
+                return true;
             });
 
             useAutoBackup = (ListPreference)findPreference(getString(R.string.settings_key_auto_backup_password_enc));
@@ -445,38 +446,38 @@ public class SettingsActivity extends BaseActivity
                 addPreferencesFromResource(R.xml.preferences_special);
 
                 Preference clearKeyStore = findPreference(getString(R.string.settings_key_clear_keystore));
-                clearKeyStore.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                    @Override
-                    public boolean onPreferenceClick(Preference preference) {
-                        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+                clearKeyStore.setOnPreferenceClickListener(preference -> {
+                    AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
 
-                        builder.setTitle(R.string.settings_dialog_title_clear_keystore);
-                        if (settings.getEncryption() == EncryptionType.PASSWORD)
-                            builder.setMessage(R.string.settings_dialog_msg_clear_keystore_password);
-                        else if (settings.getEncryption() == EncryptionType.KEYSTORE)
-                            builder.setMessage(R.string.settings_dialog_msg_clear_keystore_keystore);
+                    builder.setTitle(R.string.settings_dialog_title_clear_keystore);
+                    if (settings.getEncryption() == EncryptionType.PASSWORD)
+                        builder.setMessage(R.string.settings_dialog_msg_clear_keystore_password);
+                    else if (settings.getEncryption() == EncryptionType.KEYSTORE)
+                        builder.setMessage(R.string.settings_dialog_msg_clear_keystore_keystore);
 
-                        builder.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialogInterface, int i) {
-                                KeyStoreHelper.wipeKeys(getActivity());
-                                if (settings.getEncryption() == EncryptionType.KEYSTORE) {
-                                    DatabaseHelper.wipeDatabase(getActivity());
-                                    ((SettingsActivity) getActivity()).generateNewEncryptionKey();
-                                }
-                            }
-                        });
-                        builder.setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialogInterface, int i) {
-                            }
-                        });
+                    builder.setPositiveButton(android.R.string.ok, (dialogInterface, i) -> {
+                        KeyStoreHelper.wipeKeys(getActivity());
+                        if (settings.getEncryption() == EncryptionType.KEYSTORE) {
+                            DatabaseHelper.wipeDatabase(getActivity());
+                            ((SettingsActivity) getActivity()).generateNewEncryptionKey();
+                        }
+                    });
+                    builder.setNegativeButton(android.R.string.cancel, (dialogInterface, i) -> {
+                    });
 
-                        builder.create().show();
-
-                        return false;
-                    }
+                    builder.create().show();
+                    return false;
                 });
+            }
+            // Remove Theme Mode selection option for devices below Android 10. Disable theme selection if Theme Mode is set auto
+            if(Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+                catUI.removePreference(themeMode);
+                catUI.removePreference(themeBlack);
+            } else {
+                if (sharedPref.getString(getString(R.string.settings_key_theme_mode), getString(R.string.settings_default_theme_mode)).equals("auto"))
+                    catUI.removePreference(theme);
+                else
+                    catUI.removePreference(themeBlack);
             }
         }
     }
